refactor(sputnik): integrate rawprincipal into ic_cdk
