+++ src/sputnik/src/js/apis/ic_cdk/types.rs
@@ -1,0 +1,6 @@
+pub mod candid {
+    use rquickjs::TypedArray;
+
+    #[derive(Clone)]
+    pub struct JsRawPrincipal<'js>(pub TypedArray<'js, u8>);
+}
