+++ src/sputnik/src/js/mod.rs
@@ -3,3 +3,5 @@ pub mod constants;
 mod dev;
 pub mod module;
 pub mod runtime;
+
+pub use apis::types;
